requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astroid==3.2.4
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.1
autopep8==2.3.1
bcrypt==4.1.3
black==24.4.2
bleach==6.1.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
cssselect==1.2.0
cssutils==2.11.1
defusedxml==0.8.0rc2
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
distro==1.9.0
dj-database-url==2.2.0
Django==5.1.3
django-allauth==65.2.0
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-deep-serializer==0.1.3
django-environ==0.11.2
django-extensions==3.2.1
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==26.0.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastapi-mail==1.4.1
filelock==3.15.4
flake8==7.1.0
frozenlist==1.4.1
geoip2==5.0.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.0
greenlet==3.0.3
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
identify==2.6.0
idna==3.7
importlib_resources==6.4.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
limits==3.13.0
lxml==5.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
maxminddb==2.6.3
mccabe==0.7.0
mdurl==0.1.2
modules==1.0.0
more-itertools==10.3.0
multidict==6.0.5
mypy-extensions==1.0.0
node==1.2.2
nodeenv==1.9.1
oauthlib==3.2.2
odict==1.9.0
openai==0.28.0
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
pillow==10.4.0
pipdeptree==2.23.1
pipenv==2023.12.1
platformdirs==4.2.2
pluggy==1.5.0
plumber==1.7
pre-commit==3.7.1
premailer==3.10.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.1
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pylint==3.2.6
PyMySQL==1.1.1
pyotp==2.9.0
pyparsing==3.2.1
pytest==8.2.2
pytest-asyncio==0.23.8
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
qrcode==8.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
SQLAlchemy==2.0.31
sqlparse==0.4.3
starlette==0.37.2
stripe==10.7.0
tomli==2.0.1
tomlkit==0.13.0
tqdm==4.67.1
twilio==9.2.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2022.6
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==2.2.2
user-agents==2.2.0
uuid7==0.1.0
uvicorn==0.30.3
virtualenv==20.26.3
virtualenvwrapper-win==1.2.7
watchfiles==0.22.0
webencodings==0.5.1
websockets==12.0
whitenoise==6.7.0
wrapt==1.16.0
yarl==1.9.4
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.1
zope.event==5.0
zope.hookable==7.0
zope.interface==7.2
zope.lifecycleevent==5.1
zope.proxy==6.1
